Create README.md
Crime is a very old concept which has moved on in society from generation to generation. Nobody is safe today. Criminals are employing scientific procedures to commit crimes as science and technology evolve, and authorities have been surprised by criminals' increased knowledge as society advances. Many crimes are perpetrated for the purpose of obtaining money or property wealth, with the most common being family feuds over property allocation. It is claimed to be a widespread societal issue that has an impact on not just a society's quality of life but also its economic progress. It is focused as major issue that has a direct impact on an individual's quality of life as well as indirect societal consequences.
With the rapid increase of crimes and the types of crimes, researchers are finding a strong connection in the various features affecting the crime like the geographic features, age, gender and many more yet to come into picture. Despite the fact that crimes can happen anywhere, it is normal for criminals to focus their efforts on regions where they are most familiar and at ease.
We seek to improve societal awareness by giving an advanced approach to identifying criminal hotspots and determining the nature, location, and time of crimes committed. Furthermore, having such a system in place would assist police in taking proactive actions in preventing and solving crimes at a far faster rate by allowing for crime prediction and prevention.
Various types of crimes are discussed here and focus on their intensities in the highest city in the highest year is put forward with machine learning concepts for the societal and police benefits in solving the crime. These crimes are represented in different visualization techniques like bar graphs, pie charts to get a clear picture about the crime.

The above problem made me to go for a research about how can solving a crime case made easier. Through many documentation and cases, it came out that machine learning and data science can make the work easier and faster. The aim of this project is to make crime prediction using the features present in the dataset. The dataset is extracted from the official sites. With the help of machine learning algorithm, using python as core we can predict the type of crime which will occur in a particular area. The objective would be to train a model for prediction. The training would be done using the training data set which will be validated using the test dataset. 

ABSTRACT

The project Criminal Record Management System is a Criminal record management system that uses to record crime activities of criminals. It can be used to report crime activities. This project is mainly useful for law and enforcement agencies. The law and enforcement authority can preserve records of the criminals and search any criminal using the system. This is an application with database system in which police will keep the record of criminals who have been arrested. We have used Python, Anaconda, Jupiter, MySql and Google colab to develop this project. We also used binary search algorithm to find a criminal from database. The project’s interface is very user friendly and helpful for authorities for easier and fast access of criminal records. 
The main objective of the project on "Criminal Database Management System" is to manage the detail of Case Id, Name, Age, Sex, Identification Mark, Crime Done, etc. It manages all the information criminal and their record. 

PREDICTIVE MODELING 
Predictive modeling is the way of building a model that is capable of making predictions. The process includes a machine learning algorithm that learns certain properties from a training dataset in order to make those predictions. Pattern classification tasks can be divided into two parts, Supervised and unsupervised learning. In supervised learning, the class labels in the dataset, which is used to build the classification model, are known. In a supervised learning problem, we would know which training dataset has the particular output which will be used to train so that prediction can be made for unseen data. 

DATA PREPROCESSING
This process includes methods to remove any null values or infinite values which may affect the accuracy of the system. The main steps include Formatting, cleaning and sampling. Cleaning process is used for removal or fixing of some missing data there may be data that are incomplete. 
Sampling is the process where appropriate data are used which may reduce the running time for the algorithm. Using python, the preprocessing is done.

MODEL SELECTION
Based on the defined goal(s) (supervised or unsupervised) we have to select one of or combinations of modeling techniques. Such as 
• KNN Classification   
• Logistic Regression   
• Decision Trees   
• Random Forest   
• Support Vector Machine (SVM)   
• Bayesian methods

IMPLEMENTATION

The dataset used in this project is taken from Kaggle.com. The dataset obtained from kaggle.com is maintained and updated by the Chicago police department. 
The implementation of this project is divided into following steps –

DATA COLLECTION
Crime dataset from kaggle.com is used in CSV format. 

DATA PREPROCESSING

10k entries are present in the dataset. The null values are removed using df = df.dropna() where df is the data frame. The categorical attributes (Location, Block, Crime Type, Community Area) are converted into numeric using Label Encoder. The date attribute is splitted into new attributes like month and hour which can be used as feature for the model. 

FEATURE SELECTION

Features selection is done which can be used to build the model. The attributes used for feature selection are Block, Location, District, Community area, X coordinate , Y coordinate, Latitude , Longitude, Hour and month,